Ignore time when creating new Test Run.
When creating a new Test Run, specially via the admin pages, the default value
for the `timestamp` field was being set to a date/time value. Since I want to
be able to sort by date and build name and the time is not really important, I
have decided to not really set the time anymore.